Teleport (#13)
* Basic teleporting works
It moves so the target is the origin, NOT the person. So it feels
okay for medium->long warps, but not quite right for short warps.

* Cleaned up code

* Working but quaternion errors with y-up

* Added a z-up frame
This should make the nav stack happy, but it is untested as of yet.

* Tested and working z-up.
Have to add extra transform, which is annoying but seems to be necessary.